fix(api): add tests to register facility
Ref. metriport/metriport-internal#1785
